# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "TVE API"
  version: "1.0.0"
  title: "TVE API"
  contact:
    email: "vadim.yudkovsky@charter.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
tags:
- name: "admins"
  description: "Secured Admin-only calls"
- name: "userse"
  description: "Operations available to regular users"
schemes:
- "https"
paths:
  /environments:
    get:
      tags:
      - "users"
      summary: "gets a list of environments"
      description: "Get all environments\n"
      operationId: "getEnvironments"
      produces:
      - "application/json"
      parameters:
      - name: "searchString"
        in: "query"
        description: "pass an optional search string for looking up inventory"
        required: false
        type: "string"
      - name: "skip"
        in: "query"
        description: "number of records to skip for pagination"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "maximum number of records to return"
        required: false
        type: "integer"
        maximum: 50.0
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EnvironmentItem"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "Users"
    post:
      tags:
      - "admins"
      summary: "adds environment"
      description: "Adds environment"
      operationId: "addEnvironment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "environmentItem"
        description: "Environment item to add"
        required: false
        schema:
          $ref: "#/definitions/EnvironmentItem"
      responses:
        201:
          description: "item created"
        400:
          description: "invalid input, object invalid"
        409:
          description: "an existing item already exists"
      x-swagger-router-controller: "Admins"
definitions:
  EnvironmentItem:
    type: "object"
    required:
    - "id"
    - "manufacturer"
    - "name"
    - "releaseDate"
    properties:
      id:
        type: "string"
        format: "uuid"
        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      name:
        type: "string"
        example: "Widget Adapter"
      releaseDate:
        type: "string"
        format: "int32"
        example: "2016-08-29T09:12:33.001Z"
      manufacturer:
        $ref: "#/definitions/Manufacturer"
    example:
      releaseDate: "2016-08-29T09:12:33.001Z"
      name: "Widget Adapter"
      id: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      manufacturer:
        phone: "408-867-5309"
        name: "ACME Corporation"
        homePage: "https://www.acme-corp.com"
  Manufacturer:
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "ACME Corporation"
      homePage:
        type: "string"
        format: "url"
        example: "https://www.acme-corp.com"
      phone:
        type: "string"
        example: "408-867-5309"
    example:
      phone: "408-867-5309"
      name: "ACME Corporation"
      homePage: "https://www.acme-corp.com"